fgghdfgh

+++ DatabaseApplicationV2/resources/views/movies/movieList.blade.php
@@ -12,17 +12,18 @@
     </tr>
   </thead>
   <tbody style="background-color: grey; color: white">
-    @foreach(Session::get('item') as $key=>$item)
-    <tr>
-      <td>{{$item->name}}</td>
-      <td>{{$item->genre}}</td>
-      <td>{{$item->release_date}}</td>
-      <td><a class="btn btn-info" href="/movies/detail/{{$item->id}}" style="margin-right: 5px">View</a><a class="btn btn-danger" href="/list/delete/{{$key}}">Delete</a></td>
-    </tr>
-    @endforeach
+    @if(Session::get('item')!=NULL)
+      @foreach(Session::get('item') as $key=>$item)
+      <tr>
+        <td>{{$item->name}}</td>
+        <td>{{$item->genre}}</td>
+        <td>{{$item->release_date}}</td>
+        <td><a class="btn btn-info" href="/movies/detail/{{$item->id}}" style="margin-right: 5px">View</a><a class="btn btn-danger" href="/list/delete/{{$key}}">Delete</a></td>
+      </tr>
+      @endforeach
+    @endif
   </tbody>
 </table>
   <a class="btn btn-primary" href="/">< Back to home</a>
   <a class="btn btn-danger" href="/list/clear">Clear the list</a>
-
 @endsection
